requirements.txt
pillow
pillow-heif
tqdm
wand